// src/public-api/model/articleCategoryDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ArticleCategoryDto { 
    id?: number;
    name?: string | null;
    slug?: string | null;
    description?: string | null;
    parentCategoryId?: number | null;
    displayOrder?: number;
    isActive?: boolean;
    createdAt?: string;
    parentCategory?: ArticleCategoryDto;
    subCategories?: Array<ArticleCategoryDto> | null;
}